+++ src/main/java/com/micheck12/back/user/auth/PrincipalDetailsService.java
@@ -1,8 +1,8 @@
-package com.micheck12.back.chat.auth;
+package com.micheck12.back.user.auth;
 
-import com.micheck12.back.chat.entity.User;
+import com.micheck12.back.user.entity.User;
 import com.micheck12.back.chat.exception.CustomException;
-import com.micheck12.back.chat.repository.UserRepository;
+import com.micheck12.back.user.repository.UserRepository;
 import com.micheck12.back.chat.util.error.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
