+++ src/main/java/com/micheck12/back/user/config/JWT/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.micheck12.back.chat.config.JWT;
+package com.micheck12.back.user.config.JWT;
 
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
