+++ src/main/java/com/micheck12/back/user/config/JWT/JwtAuthenticationFilter.java
@@ -1,9 +1,7 @@
-package com.micheck12.back.chat.config.JWT;
+package com.micheck12.back.user.config.JWT;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
