+++ src/main/java/com/micheck12/back/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.micheck12.back.chat.entity;
+package com.micheck12.back.user.entity;
 
 import com.micheck12.back.common.entity.BaseTimeEntity;
 import lombok.Builder;
